+++ tests/test_pylint.py
@@ -8,6 +8,7 @@ from textwrap import dedent
 
 from click.testing import CliRunner
 
+from acsoo.main import main
 from acsoo.pylintcmd import pylintcmd
 from acsoo.tools import working_directory
 
@@ -22,7 +23,7 @@ class TestPylint(unittest.TestCase):
         with working_directory(DATA_DIR):
             res = runner.invoke(pylintcmd, [
                 '-e', 'fixme:0,manifest-required-key',
-                '--', 'odoo',
+                '-m', 'odoo',
             ])
             self.assertTrue(res.exit_code != 0)
             expected = dedent("""\
@@ -47,3 +48,38 @@ class TestPylint(unittest.TestCase):
                   fixme: 1 (expected 0)
             """)
             assert expected in res.output
+
+    def test2_config(self):
+        runner = CliRunner()
+        with working_directory(DATA_DIR):
+            res = runner.invoke(main, [
+                '-c', os.path.join(DATA_DIR, 'test_pylint2.cfg'),
+                'pylint',
+            ])
+            self.assertTrue(res.exit_code != 0)
+            expected = dedent("""\
+                messages that did not cause failure:
+                  manifest-required-key: 1
+                messages that caused failure:
+                  fixme: 1 (expected 0)
+            """)
+            assert expected in res.output
+
+    def test3(self):
+        runner = CliRunner()
+        with working_directory(DATA_DIR):
+            res = runner.invoke(pylintcmd, [
+                '--', '-d', 'fixme',
+            ])
+            self.assertTrue(res.exit_code != 0)
+            assert 'fixme: ' not in res.output
+
+    def test3_config(self):
+        runner = CliRunner()
+        with working_directory(DATA_DIR):
+            res = runner.invoke(main, [
+                '-c', os.path.join(DATA_DIR, 'test_pylint3.cfg'),
+                'pylint',
+            ])
+            self.assertTrue(res.exit_code != 0)
+            assert 'fixme: ' not in res.output